refactor: util 폴더를 lib 경로로 이동
